fix: address Tasks 7-9 code/concurrency review findings
- A: log-and-rethrow in existingLegInstrumentIds closure (was silent try?)
- B: fiatInstrument → private fiat + fiatDenomination() accessor
- C: StubMetadata → StubMetadataResolver in MoolahTests/Support/
- D: extract makeExchangeSyncEngine() shared factory to Support/
- E: restore apiKeyProvider closure-not-value rationale comment
- F: isBetterFallback docstring updated with why-not-what rationale
- G: CoinstashAssetMetadataResolver: add explicit Sendable conformance
- H: CallTracker @unchecked Sendable lock-discipline justification doc
