+++ MoolahTests/Features/Sync/SyncedAccountStoreExchangeTests.swift
@@ -88,7 +88,7 @@ struct SyncedAccountStoreExchangeTests {
             registry: registry, fiatInstrument: .AUD,
             existingLegInstrumentIds: { [] }),
           discovery: discovery),
-        metadataResolverFactory: { _ in StubMetadata([:]) }))
+        metadataResolverFactory: { _ in StubMetadataResolver([:]) }))
     return account
   }
