+++ MoolahTests/Shared/Exchange/ExchangeSyncEngineResolutionTests.swift
@@ -8,6 +8,9 @@ import Testing
 /// Thread-safe string accumulator used to assert that metadata callbacks
 /// are (or are not) invoked. Mirrors the lock-bracket pattern used in
 /// `CryptoTokenDiscoveryTestDoubles.swift`.
+///
+/// `@unchecked Sendable`: mutable state guarded by `NSLock`; lock-bracket
+/// pattern, the project convention for non-actor concurrent test stubs.
 final class CallTracker: @unchecked Sendable {
   private let lock = NSLock()
   private var calls: [String] = []
@@ -20,50 +23,10 @@ final class CallTracker: @unchecked Sendable {
   var all: [String] { lock.withLock { calls } }
 }
 
-// MARK: - Stub metadata resolver (internal so ExchangeSyncEngineTests can reuse)
-
-final class StubMetadata: ExchangeAssetMetadataResolving, @unchecked Sendable {
-  let map: [String: ExchangeAssetMetadata?]
-  let onCall: @Sendable (String) -> Void
-
-  init(
-    _ map: [String: ExchangeAssetMetadata?],
-    onCall: @escaping @Sendable (String) -> Void = { _ in }
-  ) {
-    self.map = map
-    self.onCall = onCall
-  }
-  func assetMetadata(forSymbol symbol: String) async throws -> ExchangeAssetMetadata? {
-    onCall(symbol)
-    guard let hit = map[symbol] else { return nil }
-    return hit
-  }
-}
-
 // MARK: - Test suite
 
 @Suite("ExchangeSyncEngine resolution")
 struct ExchangeSyncEngineResolutionTests {
-  private func makeEngine(
-    registry: StubInstrumentRegistry,
-    regResolver: CountingRegistrationResolver? = nil
-  ) -> ExchangeSyncEngine {
-    let resolverToUse: CountingRegistrationResolver
-    if let regResolver {
-      resolverToUse = regResolver
-    } else {
-      let defaultResolver = CountingRegistrationResolver()
-      defaultResolver.setDefault(.success(coingecko: "id", cryptocompare: nil, binance: nil))
-      resolverToUse = defaultResolver
-    }
-    let discovery = CryptoTokenDiscoveryService(
-      registry: registry, resolver: resolverToUse, alchemy: CountingAlchemyClientStub())
-    return ExchangeSyncEngine(
-      resolver: ExchangeInstrumentResolver(
-        registry: registry, fiatInstrument: .AUD,
-        existingLegInstrumentIds: { [] }),
-      discovery: discovery)
-  }
 
   private func depositRow(_ symbol: String, _ amount: Decimal)
     -> ExchangeImportedTransaction
@@ -84,7 +47,7 @@ struct ExchangeSyncEngineResolutionTests {
   @Test
   func opDepositResolvesToRealOptimismOPNotSpam() async throws {
     let registry = StubInstrumentRegistry()
-    let meta = StubMetadata([
+    let meta = StubMetadataResolver([
       "OP": ExchangeAssetMetadata(
         symbol: "OP", name: "Optimism",
         chains: [
@@ -94,7 +57,7 @@ struct ExchangeSyncEngineResolutionTests {
             decimals: 18)
         ])
     ])
-    let result = try await makeEngine(registry: registry).build(
+    let result = try await makeExchangeSyncEngine(registry: registry).build(
       account: account(), imported: [depositRow("OP", 40167)], metadata: meta)
     let leg = try #require(result.candidates.first?.transaction.legs.first)
     #expect(leg.instrument.id == "10:0x4200000000000000000000000000000000000042")
@@ -106,8 +69,8 @@ struct ExchangeSyncEngineResolutionTests {
     let registry = StubInstrumentRegistry(
       cryptoRegistrations: CryptoRegistration.builtInPresets)
     let tracker = CallTracker()
-    let meta = StubMetadata([:], onCall: { tracker.append($0) })
-    let result = try await makeEngine(registry: registry).build(
+    let meta = StubMetadataResolver([:], onCall: { tracker.append($0) })
+    let result = try await makeExchangeSyncEngine(registry: registry).build(
       account: account(), imported: [depositRow("BTC", 1)], metadata: meta)
     let leg = try #require(result.candidates.first?.transaction.legs.first)
     #expect(leg.instrument.id == "0:native")
@@ -118,7 +81,7 @@ struct ExchangeSyncEngineResolutionTests {
   @Test
   func multiChainPicksEthereumCanonical() async throws {
     let registry = StubInstrumentRegistry()
-    let meta = StubMetadata([
+    let meta = StubMetadataResolver([
       "USDC": ExchangeAssetMetadata(
         symbol: "USDC", name: "USDC",
         chains: [
@@ -132,7 +95,7 @@ struct ExchangeSyncEngineResolutionTests {
             decimals: 6),
         ])
     ])
-    let result = try await makeEngine(registry: registry).build(
+    let result = try await makeExchangeSyncEngine(registry: registry).build(
       account: account(), imported: [depositRow("USDC", 100)], metadata: meta)
     let leg = try #require(result.candidates.first?.transaction.legs.first)
     #expect(leg.instrument.id == "1:0xa0b86991c6218b36c1d19d4a2e9eb0ce3606eb48")
@@ -141,12 +104,12 @@ struct ExchangeSyncEngineResolutionTests {
   @Test
   func nativeContractNilBuildsNativeInstrument() async throws {
     let registry = StubInstrumentRegistry()
-    let meta = StubMetadata([
+    let meta = StubMetadataResolver([
       "ETH": ExchangeAssetMetadata(
         symbol: "ETH", name: "Ethereum",
         chains: [ExchangeAssetChain(chainId: 1, contractAddress: nil, decimals: 18)])
     ])
-    let result = try await makeEngine(registry: registry).build(
+    let result = try await makeExchangeSyncEngine(registry: registry).build(
       account: account(), imported: [depositRow("ETH", 2)], metadata: meta)
     let leg = try #require(result.candidates.first?.transaction.legs.first)
     #expect(leg.instrument.id == "1:native")
@@ -163,10 +126,10 @@ struct ExchangeSyncEngineResolutionTests {
         cryptocompareSymbol: nil, binanceSymbol: nil))
     let registry = StubInstrumentRegistry(
       instruments: [real.instrument], cryptoRegistrations: [real])
-    let meta = StubMetadata([
+    let meta = StubMetadataResolver([
       "SOL": ExchangeAssetMetadata(symbol: "SOL", name: "Solana", chains: [])
     ])
-    let result = try await makeEngine(registry: registry).build(
+    let result = try await makeExchangeSyncEngine(registry: registry).build(
       account: account(), imported: [depositRow("SOL", 5)], metadata: meta)
     let leg = try #require(result.candidates.first?.transaction.legs.first)
     #expect(leg.instrument.id == "1399:native")
@@ -182,7 +145,7 @@ struct ExchangeSyncEngineResolutionTests {
     }
     let registry = StubInstrumentRegistry()
     await #expect(throws: Boom.self) {
-      _ = try await makeEngine(registry: registry).build(
+      _ = try await makeExchangeSyncEngine(registry: registry).build(
         account: account(), imported: [depositRow("OP", 1)], metadata: Throwing())
     }
   }
@@ -191,12 +154,12 @@ struct ExchangeSyncEngineResolutionTests {
   func fiatLegSkipsMetadata() async throws {
     let registry = StubInstrumentRegistry()
     let tracker = CallTracker()
-    let meta = StubMetadata([:], onCall: { tracker.append($0) })
+    let meta = StubMetadataResolver([:], onCall: { tracker.append($0) })
     let row = ExchangeImportedTransaction(
       externalId: "f1", occurredAt: Date(timeIntervalSince1970: 1_762_000_000),
       category: "DEPOSIT", direction: .credit, assetSymbol: "AUD",
       amount: 50, isFiat: true, orderId: nil)
-    let result = try await makeEngine(registry: registry).build(
+    let result = try await makeExchangeSyncEngine(registry: registry).build(
       account: account(), imported: [row], metadata: meta)
     let leg = try #require(result.candidates.first?.transaction.legs.first)
     #expect(leg.instrument == .AUD)
