+++ Shared/Exchange/CoinstashAssetMetadataResolver.swift
@@ -4,7 +4,7 @@ import Foundation
 /// provider-neutral `ExchangeAssetMetadataResolving` seam carries no
 /// token. Constructed per sync by `CoinstashSyncSource` once the
 /// account's token is read from the keychain.
-struct CoinstashAssetMetadataResolver: ExchangeAssetMetadataResolving {
+struct CoinstashAssetMetadataResolver: ExchangeAssetMetadataResolving, Sendable {
   private let client: CoinstashClient
   private let token: String
